requirements.txt
asgiref
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.5
django-allauth==0.33.0
django-ckeditor==5.3.0
django-cleanup==1.0.1
django-decouple==2.1
django-haystack==2.6.1
django-js-asset==0.1.1
django-taggit==0.23.0
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.1
djangorestframework==3.8.2
elasticsearch==1.9.0
idna==2.6
oauthlib==2.0.3
olefile==0.44
Pillow
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.11.0
Unidecode==1.0.22
Unipath==1.1
urllib3==1.22
whitenoise
